requirements.txt
astroid==2.1.0
Django==2.1.3
django-bootstrap3==11.0.0
django-bootstrap4==0.0.7
django-debug-toolbar==1.10.1
django-disqus==0.5
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mock==2.0.0
pbr==5.1.1
Pillow==5.3.0
pkg-resources==0.0.0
psycopg2-binary==2.7.6.1
pylint==2.2.2
pytz==2018.7
six==1.11.0
sqlparse==0.2.4
typed-ast==1.1.0
wrapt==1.10.11
